Update implementation to use new `Scenarios`
This commit updates this formatter to use the new `Scenario` structs in
Benchee core. There's probably some additional refactoring that could
be done, but it's such a small library as it is I didn't want to
over-do it with optimizations at this point.

I figured I'd start with an easy one before doing CSV and HTML :wink: